+++ workspace/include/Judge/AngleDecision.h
@@ -9,13 +9,13 @@ class AngleDecision : public UseJudgement {
 public:
 
 	AngleDecision();
-	AngleDecision( float target_angle, int8_t range );
+	AngleDecision( double target_angle, int8_t range );
 
 	int8_t decide( void );//角度判定メソッド
 
 private:
 
-	float m_target_angle;//目標角度
+	double m_target_angle;//目標角度
 	uint8_t m_range;//使用範囲
 };
