+++ workspace/src/Action/LineTrace.cpp
@@ -10,7 +10,7 @@ LineTrace::LineTrace()
 
 /// コンストラクタ引数あり
 /// メンバー変数に格納
-LineTrace::LineTrace(GAIN _gain,float _target_val,int8_t _edge,COORDINATE _coordinate)
+LineTrace::LineTrace(GAIN _gain,double _target_val,int8_t _edge,COORDINATE _coordinate)
 {
 	gain = _gain;
 	target_val = _target_val;
@@ -44,10 +44,10 @@ int8_t LineTrace::run(int32_t speed)
 	HSV_DATA 	hsv_data;
 	COORDINATE 	current_coordinate;
 	MOTOR_POWER motor_power;
-	float 		deviation 		= 0.0f;
-	float 		pid_revision 	= 0.0f;
-	float		distance		= 0.0f; 
-	float		motor_revision  = 0.0f;
+	double 		deviation 		= 0.0f;
+	double 		pid_revision 	= 0.0f;
+	double		distance		= 0.0f; 
+	double		motor_revision  = 0.0f;
 	
 	// 色空間からhsv値を取得してから、閾値と取得した値との差分をpid制御へ渡し、操作量を取得する
 	hsv_data 		= color_space.getHSV();
